move `section-wrapper` to `shared` folder
