+++ packages/hash/frontend/src/pages/[account-slug]/shared/section-empty-state.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from "@mui/material";
 import { ReactNode } from "react";
-import { WhiteCard } from "../../../../shared/white-card";
+import { WhiteCard } from "./white-card";
 
 interface SectionEmptyStateProps {
   title: string;
